section comment 258

+++ app/Http/Controllers/AdminPostsController.php
@@ -63,8 +63,15 @@ class AdminPostsController extends Controller
     public function destroy($id)
     {
         $post = Post::findOrFail($id);
-        unlink(public_path() . $post->photo->file);
+        if(isset($post->photo->file)) {
+            unlink(public_path() . $post->photo->file);
+        }
         $post->delete();
         return redirect('/admin/posts');
+    }
+    public function post($id)
+    {
+       $post = Post::findOrFail($id);
+       return view('post', compact('post'));
     }
 }
