+++ app/Http/Controllers/AdminUsersController.php
@@ -71,7 +71,9 @@ class AdminUsersController extends Controller
     public function destroy($id)
     {
         $user = User::findOrFail($id);
-        unlink(public_path() . $user->photo->file );
+         if(isset($user->photo->file)) {
+             unlink(public_path() . $user->photo->file );
+         }
         $user->delete();
         Session::flash('deleted_user', 'The user has been deleted');
         return redirect('/admin/users');
